Moved sendmail from mailing_partner to task module
Updated CI workflow

Updated Dockerfile
