+++ docker/Dockerfile-dev
@@ -3,7 +3,7 @@ FROM condaforge/mambaforge
 RUN apt-get -qq update --yes \
   && apt-get -qq install --yes --no-install-recommends \
   build-essential git make postgresql-client \
-  ca-certificates wget locales cron \
+  ca-certificates wget locales cron sudo \
   && rm -rf /var/lib/apt/lists/*
 
 #Set locale
@@ -17,7 +17,6 @@ COPY environment-dev.yml /tmp/environment.yml
 
 # Use environment to update the env base
 RUN conda update -n base -c defaults conda \
-  && conda install mamba -n base -c conda-forge \
   && mamba update --all --yes --quiet \
   && mamba env update --file /tmp/environment.yml --name base \
   && mamba clean -afy
@@ -27,20 +26,20 @@ ARG HOST_UID
 ARG HOST_GID
 COPY docker/prepare_permission.sh /prepare_permission.sh
 RUN /prepare_permission.sh
-
 ## Copy code to the container with permissions to user
 COPY --chown=deploy:deploy docker/test.sh  /srv/deploy/docker/test.sh
 COPY --chown=deploy:deploy AlertaDengue  /srv/deploy/AlertaDengue
 COPY --chown=deploy pyproject.toml setup.cfg setup.py LICENSE README.md HISTORY.md /srv/deploy/
 COPY --chown=deploy requirements-dev.txt requirements.txt /srv/deploy/
-COPY --chown=deploy example_env_file /srv/deploy/.env_staging
+COPY --chown=deploy .env.tmpl /srv/deploy/
 
 RUN mkdir /srv/deploy/logs \
   && chown -R deploy:deploy /srv/deploy/logs
 
 # Create script activate environment
-RUN echo 'source /opt/conda/bin/activate base && exec "$@"' > activate.sh
-RUN chmod +x /activate.sh
+RUN echo 'source /opt/conda/bin/activate base && exec "$@"' > activate.sh \
+&& chmod +x /activate.sh
+
 WORKDIR /srv/deploy/AlertaDengue
 
 # Configure cron for DBF request
@@ -61,4 +60,4 @@ ENV PATH /srv/deploy/.local/bin:$PATH
 RUN cd ..; pip install --user -e .
 
 ENTRYPOINT ["bash", "/activate.sh"]
-CMD ["/entrypoint.sh", "/srv/deploy/AlertaDengue/runwsgi.sh"]
+CMD ["/entrypoint.sh"]
