+++ docker/docker-compose.yml
@@ -17,7 +17,7 @@ services:
       context: ".."
       dockerfile: "docker/Dockerfile"
       args:
-        PYTHON_VERSION: ${PYTHON_VERSION:-3.8}
+        PYTHON_VERSION: ${PYTHON_VERSION:-3.9}
         HOST_UID: ${HOST_UID}
         HOST_GID: ${HOST_GID}
     env_file:
@@ -46,7 +46,7 @@ services:
       context: ".."
       dockerfile: "docker/Dockerfile-celery"
       args:
-        PYTHON_VERSION: ${PYTHON_VERSION:-3.8}
+        PYTHON_VERSION: ${PYTHON_VERSION:-3.9}
         HOST_UID: ${HOST_UID}
         HOST_GID: ${HOST_GID}
     env_file:
