+++ docker/entrypoint.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
-echo "Executing the entrypoint!"
-echo "Install python packages..."
-envsubst < /srv/deploy/.env.tmpl > /srv/deploy/.env \
-&& pip install .
+
+echo "Create environment variables..."
+envsubst < /srv/deploy/.env.tmpl > /srv/deploy/.env
 echo "Start crontab..."
-sudo service cron start & tail -f /var/log/cron.log
+sudo service cron start
+echo "Start django..."
+bash /srv/deploy/AlertaDengue/runwsgi.sh
